rename docker to docker_util to avoid shadowing the docker lib when executing the battle script from the algobattle folder directly
