+++ algobattle/fight_handler.py
@@ -3,7 +3,7 @@ import logging
 from configparser import ConfigParser
 from typing import Any, Tuple
 
-from algobattle.docker import DockerError
+from algobattle.docker_util import DockerError
 from algobattle.team import Matchup, Team
 from algobattle.problem import Problem
